Indents, Move menu clickable&works, HideAll menus
Indentation is the key to organizing code bases or library.

The 'move' button/menu display now can be clicked on if you want to move
the 'source element' around and not just in a specific direction, also
the 'move' menu is now functionable.

'HideAll' function was created to be called upon every action that takes
place in the menu.